% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misterigauss.R
\name{misterigauss}
\alias{misterigauss}
\title{MR MiSTERI for a continous outcome with Gaussian errors.}
\usage{
misterigauss(Z = Z, A = A, Y = Y)
}
\arguments{
\item{Z}{an IV scalar variable}

\item{A}{the exposure variable}

\item{Y}{the continuous outcome variable}
}
\value{
a list object that contains causal effect estimates and standard errors.
}
\description{
This function estimates the causal effect of treatment on the treated (ETT) for a continous outcome with Gaussian error terms.
}
\references{
https://www.medrxiv.org/content/10.1101/2020.09.29.20204420v3
}
