% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misterigaussmix.R
\name{misterigaussmix}
\alias{misterigaussmix}
\title{MR MiSTERI for a continous outcome with Gaussian mixture errors.}
\usage{
misterigaussmix(Z, A, Y, maxiter = 100, tol = 0.001)
}
\arguments{
\item{Z}{an IV scalar variable}

\item{A}{the exposure variable}

\item{Y}{the continuous outcome variable}
}
\value{
a list object that contains causal effect estimates and standard errors.
}
\description{
This function estimates the causal effect of treatment on the treated (ETT) for a continous outcome with  error terms that follow Gaussian mixture distributions.
}
\references{
https://www.medrxiv.org/content/10.1101/2020.09.29.20204420v3
#' @import alabama
}
