% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misterimawii.R
\name{misterimawii}
\alias{misterimawii}
\title{MR MiSTERI with many weak invalid IVs}
\usage{
misterimawii(Z, A, Y)
}
\arguments{
\item{Z}{an IV matrix with columns representing IVs}

\item{A}{the exposure variable}

\item{Y}{the continuous outcome variable}
}
\value{
a list object that contains causal effect estimates and standard errors.
}
\description{
misterimawii combines many weak invalid IVs to reduce weak IV bias.
}
\references{
https://www.medrxiv.org/content/10.1101/2020.09.29.20204420v3
}
